groundwork for collection
