+++ shoebox/app/com/keepit/dev/DevModule.scala
@@ -17,7 +17,7 @@ import com.keepit.common.mail._
 import com.keepit.inject._
 import com.keepit.model.{PhraseRepo, BookmarkRepo, NormalizedURIRepo}
 import com.keepit.search.{ArticleStore, ResultClickTracker}
-import com.keepit.search.graph.{URIGraph, URIGraphImpl, URIGraphFields}
+import com.keepit.search.graph.{URIGraph, URIGraphImpl, URIGraphFields, URIGraphIndexer}
 import com.keepit.search.index.{ArticleIndexer, DefaultAnalyzer}
 import com.keepit.search.phrasedetector.PhraseIndexer
 import com.keepit.search.query.parser.{FakeSpellCorrector, SpellCorrector}
@@ -172,12 +172,15 @@ class SearchDevModule extends ScalaModule with Logging {
 
   @Singleton
   @Provides
-  def uriGraph(bookmarkRepo: BookmarkRepo,
-    db: Database): URIGraph = {
+  def uriGraph(bookmarkRepo: BookmarkRepo, db: Database): URIGraph = {
+    new URIGraphImpl(uriGraphIndexer(bookmarkRepo, db))
+  }
+
+  private def uriGraphIndexer(bookmarkRepo: BookmarkRepo, db: Database): URIGraphIndexer = {
     val dir = getDirectory(current.configuration.getString("index.urigraph.directory"))
     log.info(s"storing URIGraph in $dir")
     val config = new IndexWriterConfig(Version.LUCENE_41, DefaultAnalyzer.forIndexing)
-    new URIGraphImpl(dir, config, URIGraphFields.decoders(), bookmarkRepo, db)
+    new URIGraphIndexer(dir, config, URIGraphFields.decoders(), bookmarkRepo, db)
   }
 
   @Singleton
