+++ shoebox/test/com/keepit/search/MainSearcherTest.scala
@@ -29,7 +29,7 @@ import com.keepit.common.service.FortyTwoServices
 import org.apache.lucene.index.IndexWriterConfig
 import org.apache.lucene.store.{Directory, MMapDirectory, RAMDirectory}
 import org.apache.lucene.util.Version
-import com.keepit.search.graph.{URIGraph, URIGraphImpl, URIGraphFields}
+import com.keepit.search.graph.{URIGraph, URIGraphImpl, URIGraphFields, URIGraphIndexer}
 import org.apache.lucene.util.Version
 import com.keepit.shoebox.ShoeboxServiceClient
 
@@ -46,7 +46,7 @@ class MainSearcherTest extends Specification with DbRepos {
   def initIndexes(store: ArticleStore) = {
     val config = new IndexWriterConfig(Version.LUCENE_41, DefaultAnalyzer.forIndexing)
     val articleIndexer = new ArticleIndexer(new RAMDirectory, config, store, db, inject[NormalizedURIRepo], null, inject[ShoeboxServiceClient])
-    val uriGraph = new URIGraphImpl(new RAMDirectory, config, URIGraphFields.decoders(), bookmarkRepo, db)
+    val uriGraph = new URIGraphImpl(new URIGraphIndexer(new RAMDirectory, config, URIGraphFields.decoders(), bookmarkRepo, db))
     implicit val clock = inject[Clock]
     implicit val fortyTwoServices = inject[FortyTwoServices]
     val mainSearcherFactory = new MainSearcherFactory(
